+++ neoforge/src/main/java/io/github/mortuusars/envelope/neoforge/RegisterImpl.java
@@ -33,6 +33,7 @@ import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.levelgen.feature.Feature;
 import net.minecraft.world.level.levelgen.feature.configurations.FeatureConfiguration;
 import net.neoforged.neoforge.common.extensions.IMenuTypeExtension;
+import net.neoforged.neoforge.registries.DeferredHolder;
 import net.neoforged.neoforge.registries.DeferredRegister;
 import net.neoforged.neoforge.registries.NeoForgeRegistries;
 
@@ -101,8 +102,8 @@ public class RegisterImpl {
         });
     }
 
-    public static void entityDataSerializer(String id, EntityDataSerializer<?> serializer) {
-        ENTITY_DATA_SERIALIZERS.register(id, () -> serializer);
+    public static <T> Supplier<EntityDataSerializer<T>> entityDataSerializer(String id, EntityDataSerializer<T> serializer) {
+        return ENTITY_DATA_SERIALIZERS.register(id, () -> serializer);
     }
 
     public static <T extends SoundEvent> Supplier<T> soundEvent(String id, Supplier<T> supplier) {