+++ neoforge/src/main/java/io/github/mortuusars/envelope/neoforge/datagen/BuiltInDatapackEntries.java
@@ -1,6 +1,7 @@
 package io.github.mortuusars.envelope.neoforge.datagen;
 
 import io.github.mortuusars.envelope.Envelope;
+import io.github.mortuusars.envelope.world.entity.PigeonVariant;
 import io.github.mortuusars.envelope.world.item.component.seal.SealImpression;
 import io.github.mortuusars.envelope.world.item.component.seal.SealMaterial;
 import io.github.mortuusars.envelope.world.mail.service.ServiceAddresses;
@@ -14,6 +15,7 @@ import java.util.concurrent.CompletableFuture;
 
 public class BuiltInDatapackEntries extends DatapackBuiltinEntriesProvider {
     public static final RegistrySetBuilder REGISTRIES = new RegistrySetBuilder()
+          .add(Envelope.Registries.PIGEON_VARIANT, PigeonVariant::bootstrap)
           .add(Envelope.Registries.SERVICE_ADDRESS_DEFINITION, ServiceAddresses::bootstrap)
           .add(Envelope.Registries.SEAL_MATERIAL, SealMaterial::bootstrap)
           .add(Envelope.Registries.SEAL_IMPRESSION, SealImpression::bootstrap);